+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 2.1.0 (May 16, 2018)
+* Add comment action
+* Add support for relative match patterns (resolve from executing directory)
+* Fix remove action only removing first match
+
 ## 2.0.0 (May 13, 2018)
 * Refactor YAML configuration to single document (enables anchors & aliases)
 * Add node filter command line argument